refactor: split createSessionFacade god function into phase modules
The 955-line factory mixed model resolution, store setup, prompt
construction, tool state, extension wiring, and LLM client building in
one closure with four mutable variables. Split into src/core/facade/:

- model-resolution.ts: initial model + thinking level
- store-setup.ts:      store, workspace lock, crash recovery, fork/resume
- fork.ts:             prepareForkedSession (fork prep isolated for the
                       upcoming zero-copy cross-workspace fork work)
- prompt-builder.ts:   system prompt construction (pure, DI'd)
- tool-assembly.ts:    ToolAssembly class owning available/active tool
                       state + prompt refresh (replaces closure vars)
- extension-binding.ts: ExtensionRunner core API wiring
- llm-client.ts:       LLMClient factory
- adaptive-thinking.ts: adaptive-thinking fallback (re-exported)

createSessionFacade is now a ~250-line orchestrator that sequences the
phases. Public API (options/result types, setSchedulerEngine/
setLlmClient escape hatches) unchanged. No behavior change.
